Moved file classes to Storage namespace.
